+++ CHANGELOG.md
@@ -12,6 +12,7 @@ the [PEP 440 version scheme](https://peps.python.org/pep-0440/#version-scheme).
 
 ### Changed
 - AbstractPlugin to NamedPluginMixin
+- MultiprocesingPlugin's configure_logging parameter to configure_child_process_logger
 
 ### Fixed
 - The Alpha status classifier