+++ CHANGELOG.md
@@ -14,6 +14,8 @@ the [PEP 440 version scheme](https://peps.python.org/pep-0440/#version-scheme).
 ### Changed
 - AbstractPlugin to NamedPluginMixin
 - MultiprocesingPlugin's configure_logging parameter to configure_child_process_logger
+- PluginLoader returns a MultiUsePlugin
+- MultiprocessingPlugin is a SingleUsePlugin
 
 ### Fixed
 - The Alpha status classifier